+++ src/services/ICertificatePayloadGenerator.ts
@@ -1,0 +1,5 @@
+import { ITestResult } from '../models/ITestResult';
+
+export interface ICertificatePayloadGenerator {
+  generate(testResult: ITestResult): any;
+}
